feat(shared-core): Add comprehensive reporting service with multi-format export
- Create ReportDTOs.cs with request/response DTOs for sales, inventory, and financial reports
- Implement IReportService interface for report generation and management
- Add ReportService with support for CSV, PDF, and Excel export formats
- Enhance DashboardService and DashboardDataAggregationService for report integration
- Add ReportDataAccuracyPropertyTest for data validation and consistency
- Remove deprecated settings classes (BusinessSettings, TaxSettings, CurrencySettings, LocalizationSettings) from BusinessManagementDTOs.cs
- Support multi-tenant reporting with business-wide and shop-specific report generation
- Enable advanced filtering, grouping, and breakdown analysis for sales, inventory, and financial data

+++ src/Shared.Core/Services/DashboardService.cs
@@ -4,6 +4,7 @@ using Shared.Core.Entities;
 using Shared.Core.Enums;
 using Shared.Core.Repositories;
 using System.Globalization;
+using System.Linq;
 
 namespace Shared.Core.Services;
 
@@ -708,7 +709,7 @@ public class DashboardService : IDashboardService
             var todayRevenue = todaySales.Sum(s => s.TotalAmount);
             var yesterdayRevenue = yesterdaySales.Sum(s => s.TotalAmount);
             var todayTransactions = todaySales.Count;
-            var averageOrderValue = todayTransactions > 0 ? todayRevenue / todayTransactions : 0;
+            var averageOrderValue = todayTransactions > 0 ? todayRevenue / (decimal)todayTransactions : 0;
 
             var revenueGrowthPercentage = yesterdayRevenue > 0 
                 ? ((todayRevenue - yesterdayRevenue) / yesterdayRevenue) * 100 
@@ -755,7 +756,7 @@ public class DashboardService : IDashboardService
 
             var revenue = sales.Sum(s => s.TotalAmount);
             var transactionCount = sales.Count;
-            var averageOrderValue = transactionCount > 0 ? revenue / transactionCount : 0;
+            var averageOrderValue = transactionCount > 0 ? revenue / (decimal)transactionCount : 0;
             var profitEstimate = revenue * 0.3m; // Estimate
             var inventoryValue = stocks.Sum(s => s.Quantity * (s.Product?.UnitPrice ?? 0));
 
@@ -767,7 +768,7 @@ public class DashboardService : IDashboardService
                 TransactionCount = transactionCount,
                 AverageOrderValue = averageOrderValue,
                 ProfitEstimate = profitEstimate,
-                ProductCount = products.Count(),
+                ProductCount = (decimal)products.Count(),
                 InventoryValue = inventoryValue,
                 Performance = new PerformanceMetrics
                 {
